api/requirements.txt
amqp==2.2.2
Babel==2.5.3
billiard==3.5.0.3
celery==4.2.1
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10.1
kombu==4.2.0
MarkupSafe==1.0
pytz==2018.3
redis==2.10.6
tornado==5.0.2
vine==1.1.4
Werkzeug==0.14.1
Flask-BasicAuth
pyyaml
